refactor(blocks): BlockInspector を per-type ディスパッチャ＋サブエディタへ分割する (#501)
1017行の god component を解体する。(1) block.type の if ラダー（末尾が暗黙 hero）を ReactElement 返り型の exhaustive switch にし、ブロック型追加時の hero 誤フォールスルーをコンパイルエラーで防ぐ。(2) 4つのサブエディタ（HeroMediaField / GalleryItemsField / SeriesField / GroupChildrenField）を ui/inspectors/ へ分離（GroupChildrenField は BlockInspector を再帰 import）。(3) kind/variant/layout/chartType/tone/size で重複していた Select+options+label-key を EnumSelect ヘルパへ集約。ディスパッチャは 456 行に縮小。挙動不変。
